# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Scrum Poker API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/rooms": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "parameters": [
                    {
                        "description": "Create a new room",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rooms.RoomNewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rooms.RoomNewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/rooms/{pincode}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Get room informations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rooms.RoomGetResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/rooms/{pincode}/join": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Join a room",
                "parameters": [
                    {
                        "description": "Join a room",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rooms.RoomJoinRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rooms.RoomJoinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/rooms/{pincode}/players": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Get players from a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/players.Player"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/rooms/{pincode}/players/{id}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Rename player",
                "parameters": [
                    {
                        "description": "User name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rooms.RenameUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/rooms/{pincode}/topic": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Set a topic",
                "parameters": [
                    {
                        "description": "Set a topic",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rooms.RoomTopicRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/votings/{pincode}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Votings"
                ],
                "summary": "Get votes in a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/votings.Vote"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Votings"
                ],
                "summary": "Vote in a room",
                "parameters": [
                    {
                        "description": "Vote fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/votings.RegisterVoteRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/votings/{pincode}/reset": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Votings"
                ],
                "summary": "Reset votes in a room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pin Code",
                        "name": "pincode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "players.Player": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "joined_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "rooms.RenameUserRequest": {
            "type": "object",
            "properties": {
                "player_name": {
                    "type": "string"
                }
            }
        },
        "rooms.Room": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "rooms.RoomGetResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/players.Player"
                    }
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "rooms.RoomJoinRequest": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "boolean"
                },
                "player_name": {
                    "type": "string"
                }
            }
        },
        "rooms.RoomJoinResponse": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "room": {
                    "$ref": "#/definitions/rooms.Room"
                }
            }
        },
        "rooms.RoomNewRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "rooms.RoomNewResponse": {
            "type": "object",
            "properties": {
                "pincode": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                }
            }
        },
        "rooms.RoomTopicRequest": {
            "type": "object",
            "properties": {
                "topic": {
                    "type": "string"
                }
            }
        },
        "votings.RegisterVoteRequest": {
            "type": "object",
            "properties": {
                "player_id": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "votings.Vote": {
            "type": "object",
            "properties": {
                "player_id": {
                    "type": "string"
                },
                "player_name": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                },
                "voted_at": {
                    "type": "string"
                }
            }
        }
    }
}