# internal/models/error.go
package models

import "encoding/json"

type Error struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

func (err Error) ToJson() (string, error) {
	str, e := json.Marshal(err)
	return string(str), e
}